refactor: Improve packages organization
