+++ app/src/main/java/ardjomand/leonardo/nutrimeal/cart/CartFragment.java
@@ -19,6 +19,7 @@ import android.widget.Toast;
 import java.util.ArrayList;
 
 import ardjomand.leonardo.nutrimeal.R;
+import ardjomand.leonardo.nutrimeal.data.pojos.CartMeal;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
