+++ app/src/main/java/ardjomand/leonardo/nutrimeal/cart/EditCartMealContract.java
@@ -1,5 +1,7 @@
 package ardjomand.leonardo.nutrimeal.cart;
 
+import ardjomand.leonardo.nutrimeal.data.pojos.CartMeal;
+
 interface EditCartMealContract {
 
     interface View {
