+++ app/src/main/java/ardjomand/leonardo/nutrimeal/data/EditCartInteractor.java
@@ -1,6 +1,6 @@
 package ardjomand.leonardo.nutrimeal.data;
 
-import ardjomand.leonardo.nutrimeal.meals.Meal;
+import ardjomand.leonardo.nutrimeal.data.pojos.Meal;
 
 public interface EditCartInteractor {
