+++ app/src/main/java/ardjomand/leonardo/nutrimeal/data/EditCartMealQuantityInteractor.java
@@ -1,6 +1,6 @@
 package ardjomand.leonardo.nutrimeal.data;
 
-import ardjomand.leonardo.nutrimeal.cart.CartMeal;
+import ardjomand.leonardo.nutrimeal.data.pojos.CartMeal;
 
 public interface EditCartMealQuantityInteractor {
