+++ app/src/main/java/ardjomand/leonardo/nutrimeal/data/EditMealInteractor.java
@@ -2,7 +2,7 @@ package ardjomand.leonardo.nutrimeal.data;
 
 import android.net.Uri;
 
-import ardjomand.leonardo.nutrimeal.meals.Meal;
+import ardjomand.leonardo.nutrimeal.data.pojos.Meal;
 
 public interface EditMealInteractor {
