+++ app/src/main/java/ardjomand/leonardo/nutrimeal/data/common/GenericRepositoryImpl.java
@@ -1,4 +1,4 @@
-package ardjomand.leonardo.nutrimeal.data;
+package ardjomand.leonardo.nutrimeal.data.common;
 
 import android.util.Log;
 
@@ -6,6 +6,8 @@ import com.google.firebase.database.ChildEventListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
+
+import ardjomand.leonardo.nutrimeal.data.pojos.KeyClass;
 
 public class GenericRepositoryImpl<T extends KeyClass> implements GenericRepository.Repository {
