+++ app/src/main/java/ardjomand/leonardo/nutrimeal/data/pojos/KeyClass.java
@@ -1,4 +1,4 @@
-package ardjomand.leonardo.nutrimeal.data;
+package ardjomand.leonardo.nutrimeal.data.pojos;
 
 public class KeyClass {
