+++ app/src/main/java/ardjomand/leonardo/nutrimeal/editmeal/EditMealContract.java
@@ -2,7 +2,7 @@ package ardjomand.leonardo.nutrimeal.editmeal;
 
 import android.net.Uri;
 
-import ardjomand.leonardo.nutrimeal.meals.Meal;
+import ardjomand.leonardo.nutrimeal.data.pojos.Meal;
 
 interface EditMealContract {
