+++ app/src/main/java/ardjomand/leonardo/nutrimeal/meals/MealsContract.java
@@ -2,6 +2,8 @@ package ardjomand.leonardo.nutrimeal.meals;
 
 import java.util.List;
 
+import ardjomand.leonardo.nutrimeal.data.pojos.Meal;
+
 interface MealsContract {
 
     interface View {
